webapp/requirements.txt
django~=4.0.6
djangorestframework-simplejwt
django-cors-headers==3.8.0
drf-yasg==1.20.0
psycopg2
djangorestframework~=3.13.1
PyJWT~=2.4.0
django-filter
django_auto_one_to_one
Faker
phonenumbers